@@ -1,13 +1,13 @@
 buildscript {
     ext {
-        kotlin_version = '1.9.24'
+        kotlin_version = '2.1.0'
         coroutinesVersion = '1.6.4'
         ok_http_version = '4.9.1'
         dashjVersion = '21.1.7-SNAPSHOT'
         dppVersion = "1.7.3"
-        hiltVersion = '2.51'
+        hiltVersion = '2.53'
         hiltCompilerVersion = '1.2.0'
-        hiltWorkVersion = '1.0.0'
+        hiltWorkVersion = '1.2.0'
         workRuntimeVersion='2.7.1'
         firebaseVersion = '32.1.1'
         roomVersion = '2.5.2'
@@ -64,7 +64,8 @@ buildscript {
 }
 
 plugins {
-    id 'com.google.devtools.ksp' version '1.9.24-1.0.20' apply false
+    id 'com.google.devtools.ksp' version '2.1.0-1.0.29' apply false
+    id("org.jetbrains.kotlin.plugin.compose") version "$kotlin_version" apply false
 }
 
 allprojects {
